moving effortcalculator to own file
