+++ wrapper/src/api/providers/algorithm/util/effortPoints/baseEffort.ts
@@ -1,5 +1,5 @@
 import { Metric } from '../../../../server/types';
-import { LogarithmicEffort, ManhattanDistance, MultiplyEffort } from './';
+
 export abstract class BaseEffort implements Metric {
   constructor(
     protected readonly expertise: number,
@@ -13,20 +13,4 @@ export abstract class BaseEffort implements Metric {
 
   info?: string;
   infoLink?: string;
-}
-
-export function calculateEffort(expertise: number, hrAmount: number): Metric[] {
-  const efforts = [ManhattanDistance, LogarithmicEffort, MultiplyEffort];
-  return efforts
-    .map((Effort) => new Effort(expertise, hrAmount))
-    .map(({ value, formula, name, range, info, infoLink }) => {
-      return {
-        value,
-        formula,
-        name,
-        range,
-        info,
-        infoLink,
-      };
-    });
 }
