+++ wrapper/src/api/providers/algorithm/util/effortPoints/calculateEffort.ts
@@ -1,0 +1,18 @@
+import { Metric } from '../../../../server/types';
+import { LogarithmicEffort, ManhattanDistance, MultiplyEffort } from './';
+
+export function calculateEffort(expertise: number, hrAmount: number): Metric[] {
+  const efforts = [ManhattanDistance, LogarithmicEffort, MultiplyEffort];
+  return efforts
+    .map((Effort) => new Effort(expertise, hrAmount))
+    .map(({ value, formula, name, range, info, infoLink }) => {
+      return {
+        value,
+        formula,
+        name,
+        range,
+        info,
+        infoLink,
+      };
+    });
+}
